rename paypal logo icon
