+++ packages/@guardian/source-react-components/src/icons/Icons.stories.tsx
@@ -45,7 +45,7 @@ import { SvgInfoRound } from './components/SvgInfoRound';
 import { SvgLinkedIn } from './components/SvgLinkedIn';
 import { SvgMediaControlsPlay } from './components/SvgMediaControlsPlay';
 import { SvgMinus } from './components/SvgMinus';
-import { SvgPaypalBrand } from './components/SvgPaypalBrand';
+import { SvgPayPalBrand } from './components/SvgPayPalBrand';
 import { SvgPerson } from './components/SvgPerson';
 import { SvgPinned } from './components/SvgPinned';
 import { SvgPinterest } from './components/SvgPinterest';
@@ -116,7 +116,7 @@ const uiIcons = {
 const paymentIcons = {
 	SvgCreditCard,
 	SvgDirectDebit,
-	SvgPaypalBrand,
+	SvgPayPalBrand,
 };
 
 const widePaymentIcons = {
