+++ packages/@guardian/source-react-components/src/icons/components/SvgPayPalBrand.tsx
