+++ scripts/validate-dist/source-package-exports.test.ts
@@ -170,6 +170,7 @@ describe('No exports have changed in the ', () => {
 			'SvgMinus',
 			'SvgOfflineCloud',
 			'SvgPayPal',
+			'SvgPayPalBrand',
 			'SvgPerson',
 			'SvgPinned',
 			'SvgPinterest',
